+++ src/main/java/org/springframework/samples/petclinic/configuration/SecurityConfiguration.java
@@ -44,7 +44,6 @@ public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
 				.antMatchers("/teachers/new").hasAnyAuthority("dean","admin")
 				.antMatchers("/myTeachers/**").hasAnyAuthority("student","admin")
 				.antMatchers("/teachers/*").permitAll()
-				.antMatchers("/teachers/**").permitAll()
 				.antMatchers("/teachers/*/scores").permitAll()
 				.antMatchers("/teachers/*/scores/new").hasAnyAuthority("admin", "student")
 				.antMatchers("/studentsWithScore").hasAnyAuthority("admin")
@@ -62,6 +61,7 @@ public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
 				.antMatchers("/myTeachers").permitAll()
 				.antMatchers("/reports").hasAnyAuthority("teacher")
 				.antMatchers("/reports/**").hasAnyAuthority("teacher")
+				.antMatchers("/teachers/{teacherId}/studentsRated").hasAnyAuthority("admin")
 				.anyRequest().denyAll()
 				.and()
 				 	.formLogin()
@@ -100,3 +100,5 @@ public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
 	}
 	
 }
+
+
