+++ src/main/java/org/springframework/samples/petclinic/service/StudentService.java
@@ -79,7 +79,10 @@ public class StudentService {
 		return studentRepository.myTeachers(id);
 
 	}
-
+	
+	public Collection<Student> StudentsRatedATeacher(int teacherId) throws DataAccessException{
+		return studentRepository.StudentsRatedATeacher(teacherId);
+	}
 	
 
 }
