+++ src/main/java/org/springframework/samples/petclinic/web/TeacherController.java
@@ -8,14 +8,12 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.samples.petclinic.model.Score;
 import org.springframework.samples.petclinic.model.Student;
-import org.springframework.samples.petclinic.model.Subject;
 import org.springframework.samples.petclinic.model.Teacher;
 import org.springframework.samples.petclinic.model.Teachers;
 import org.springframework.samples.petclinic.service.ScoreService;
@@ -54,8 +52,8 @@ public class TeacherController {
 	public void setAllowedFields(WebDataBinder dataBinder) {
 		dataBinder.setDisallowedFields("id");
 	}
-	
-	@InitBinder("report")
+
+	@InitBinder("score")
 	public void initTeacherBinder(WebDataBinder dataBinder) {
 		dataBinder.setValidator(new ScoreValidator(scoreService, studentService, teacherService));
 	}
@@ -107,7 +105,6 @@ public class TeacherController {
 	@GetMapping(value = "/teachersFound")
 	public String processFindForm(Teacher teacher, BindingResult result, Map<String, Object> model) {
 
-
 		if (teacher.getFirstName() == null) {
 			teacher.setFirstName(""); // empty string signifies broadest possible search
 		}
@@ -117,13 +114,11 @@ public class TeacherController {
 			result.rejectValue("firstName", "notFound", "not found");
 			model.put("teachers", new Teacher());
 			return "teachers/findTeachers";
-		}
-		else if (results.size() == 1) {
+		} else if (results.size() == 1) {
 			// 1 teacher found
-			 teacher = results.iterator().next();
+			teacher = results.iterator().next();
 			return "redirect:/teachers/" + teacher.getId();
-		}
-		else {
+		} else {
 			// multiple teachers found
 			model.put("selections", results);
 			return "teachers/teachersList";
@@ -134,26 +129,8 @@ public class TeacherController {
 	public String showTeacherScoreList(@PathVariable("teacherId") int teacherId, Map<String, Object> model) {
 		Collection<Score> scores = this.teacherService.findScoresByTeacherId(teacherId);
 		model.put("scores", scores);
-		return "scores/scoresList"; 
-	} 
-	
-	@GetMapping(path="/teachers/{teacherId}/scores/delete/{scoreId}")
-	public String deleteScore(@PathVariable("teacherId") int teacherId, @PathVariable("scoreId") int scoreId, ModelMap modelMap){
-		String view = "scores/scoresList";
-		Optional<Score> score = scoreService.findScoreById(scoreId);
-		if(score.isPresent()) {
-			scoreService.delete(score.get());
-			modelMap.addAttribute("message", "Score successfully deleted!");
-			view = showTeacherScoreList(teacherId, modelMap);
-		} else {
-			modelMap.addAttribute("message", "Score not found!");
-			view = showTeacherScoreList(teacherId, modelMap);
-		}
-		return view;
+		return "scores/scoresList";
 	}
-	
-
-
 
 //	@GetMapping(value = { "/teachers/{teacherId}/scores/comments" })  
 //	public String showTeacherCommentList(Teacher teacher, Map<String, Object> model) {
@@ -166,7 +143,8 @@ public class TeacherController {
 //	}
 
 	@GetMapping(value = { "teachers/{teacherId}/scores/new" })
-	public String initCreationForm(@PathVariable int teacherId, ModelMap model){ // para crear el modelo que va a la vista.
+	public String initCreationForm(@PathVariable int teacherId, ModelMap model) { // para crear el modelo que va a la
+																					// vista.
 		Score score = new Score();
 		Teacher teacher = this.teacherService.findTeacherById(teacherId);
 		score.setTeacher(teacher);
@@ -211,13 +189,19 @@ public class TeacherController {
 			model.put("teacher", teacher);
 			model.put("score", score);
 			return "scores/createForm";
-		}else {
+		} else {
 			Score uno = this.scoreService.findScoreById(scoreId);
 			BeanUtils.copyProperties(score, uno, "id", "teacher", "student");
 			this.scoreService.saveScore(score);
-		return "redirect:/teachers/{teacherId}/scores";
+			return "redirect:/teachers/{teacherId}/scores";
 		}
 	}
 
-	
+	@GetMapping(value ="teachers/{teacherId}/studentsRated")
+	public String showStudentsRatedATeacher(@PathVariable("teacherId") int teacherId, Map<String, Object> model) {
+		Collection<Student> students = this.studentService.StudentsRatedATeacher(teacherId);
+		model.put("students", students);
+		return "students/studentRatedATeacher";
+	}
+
 }
