+++ src/main/webapp/WEB-INF/jsp/scores/createForm.jsp
@@ -29,7 +29,7 @@
                         <c:when test="${score['new']}">
                             <button class="btn btn-default" type="submit">Add Score</button>
                             <h2>
-                            You must keep in mind that if you already have a score on this teacher, the new score won't be saved.
+                            *You must keep in mind that if you already have a score on this teacher, the new score won't be saved.
                             </h2>
                         </c:when>
                         <c:otherwise>
