+++ src/main/webapp/WEB-INF/jsp/subjects/newSubject.jsp
@@ -6,18 +6,16 @@
 <%@ taglib prefix="form" uri="http://www.springframework.org/tags/form" %>
 <%@ taglib prefix="petclinic" tagdir="/WEB-INF/tags" %>
 
-<petclinic:layout pageName="newSubject">
-    <h2><c:if test="${student['new']}">New </c:if> Subject</h2>
-    <form:form modelAttribute="subjects" class="form-horizontal" id="add-subject-form" action ="/subjects/save">
+<petclinic:layout pageName="subjects">
+    <h2>New Subject</h2>
+    <form:form modelAttribute="subjects" class="form-horizontal" id="add-subject-form">
         <div class="form-group has-feedback">
             <petclinic:inputField label="Name" name="name"/>
-            <petclinic:inputField label="Curso" name="curso"/>
+            <petclinic:inputField label="Course" name="course"/>
         </div>
         <div class="form-group">
             <div class="col-sm-offset-2 col-sm-10">
-            	<button class="btn btn-default" type="submit">Save Subject</button> 
-            	
-            	
+            	<button class="btn btn-default" type="submit">Add Subject</button>
                   
             </div>
         </div>
