+++ src/test/java/org/springframework/samples/petclinic/service/ReportServiceTest.java
@@ -35,16 +35,6 @@ public class ReportServiceTest {
 	
 	@Autowired
 	protected ScoreService scoreService; 
-
-	
-	ScoreRepository scoreRepo = mock(ScoreRepository.class);
-	
-	ScoreService scoreService2 = new ScoreService(scoreRepo);
-	
-	ReportRepository reportRepo = mock(ReportRepository.class);
-	
-	ReportService reportService2 = new ReportService(reportRepo);
-
 	
 	//Casos positivos
 	@Test
