+++ src/test/java/org/springframework/samples/petclinic/service/ScoreServiceTests.java
@@ -16,7 +16,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import java.util.Collection;
-import java.util.Optional;
+
 
 import org.assertj.core.util.Lists;
 import org.junit.jupiter.api.Test;
@@ -32,24 +32,15 @@ public class ScoreServiceTests {
 	
 	ScoreRepository repo = mock(ScoreRepository.class);
 	
-	ScoreService scoreService = new ScoreService(repo);
+	ScoreService scoreService = new ScoreService(repo, null);
 	
 	Score score1;
 	Score score2;
 	
-
-	
-	
-//	@BeforeEach
-//	void initAll() {
-//		score1 = new Score(7, "buen profesor", new Teacher());
-//	}
-
 	@BeforeEach
 	void initAll() {
 		score1 = new Score(7, "buen profesor",new Student(), new Teacher());
 	}
-
 	
 	//Casos positivos
 	@Test
