+++ src/test/java/org/springframework/samples/petclinic/service/SubjectServiceTest.java
@@ -34,23 +34,6 @@ public class SubjectServiceTest {
 			
 
 		}
-
-		
-		
-		
-//		@Test
-//		@DisplayName("Finding subject by id")
-//		void testFindStudentById(){
-//			
-//			Subject subject = this.subjectService.findSubjectById(111);
-//			System.out.println(subject.getName());
-//			assertThat(subject.getName()).startsWith("Diseño y pruebas");
-//		}
-//		
-		
-
-
-
 		@Test
 		@DisplayName("Finding subject by id")
 		void testFindStudentById(){
@@ -70,4 +53,3 @@ public class SubjectServiceTest {
 		
 
 }
-
