+++ src/test/java/org/springframework/samples/petclinic/service/TeacherServiceTest.java
@@ -3,128 +3,145 @@ package org.springframework.samples.petclinic.service;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.Assert.assertTrue;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
-
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.HashSet;
 import java.util.List;
+
 import java.util.Set;
 
 import org.assertj.core.util.Lists;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.ComponentScan;
+import org.springframework.dao.DataAccessException;
 import org.springframework.samples.petclinic.model.Department;
 import org.springframework.samples.petclinic.model.Score;
 import org.springframework.samples.petclinic.model.Student;
-import org.springframework.dao.DataAccessException;
-import org.springframework.samples.petclinic.model.Authorities;
-
 import org.springframework.samples.petclinic.model.Teacher;
 import org.springframework.samples.petclinic.model.User;
-import org.springframework.samples.petclinic.repository.AuthoritiesRepository;
-import org.springframework.samples.petclinic.repository.DepartmentRepository;
-import org.springframework.samples.petclinic.repository.ScoreRepository;
-import org.springframework.samples.petclinic.repository.StudentRepository;
-import org.springframework.samples.petclinic.repository.TeacherRepository;
-import org.springframework.samples.petclinic.repository.UserRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 
 @DataJpaTest(includeFilters = @ComponentScan.Filter(Service.class))
 public class TeacherServiceTest {
-
+	
 	@Autowired
-	//protected TeacherService teacherService; // = new TeacherService(repo);
+	protected StudentService studentService; 
 	
+	@Autowired
+	protected TeacherService teacherService; 
 	
-	ScoreRepository scoreRepo = mock(ScoreRepository.class);
-	
-	ScoreService scoreService = new ScoreService(scoreRepo);
-	
-	UserRepository userRepo = mock(UserRepository.class);
-	
-	UserService userService = new UserService(userRepo);
-	
-	AuthoritiesRepository authoritiesRepo = mock(AuthoritiesRepository.class);
-	
-	 AuthoritiesService authoritiesService = new AuthoritiesService(authoritiesRepo, userService);
-	
-	TeacherRepository teacherRepo = mock(TeacherRepository.class);
-	
-	TeacherService teacherService = new TeacherService(teacherRepo, userService, authoritiesService);
-	
-	StudentRepository studentRepo = mock(StudentRepository.class);
-	
-	StudentService studentService = new StudentService(studentRepo, teacherRepo);
-	
-	DepartmentRepository departmentRepo = mock(DepartmentRepository.class);
-	
-	DepartmentService departmentService = new DepartmentService(departmentRepo);
-	
-	
-	Teacher t1;
-	Teacher t2;
-	
-	// Casos positivos
-	@BeforeEach
-	void initAll() {
-		t1 = new Teacher();
-		t1.setName("Perico");
-		t1.setLastName("Pedrolo");
-		t1.setUser(new User());
-		
-		Collection<Department> departments = new ArrayList<Department>();
-		
-		Department department1 = new Department();
-		department1.setName("Math");
-		departments.add(department1);
-		
-		t1.setDepartments(departments);
-	}
+	@Autowired
+	protected DepartmentService departmentService; 
 	
 	@Test
-	@DisplayName("Finding all teachers")
-	void testFindAll() throws DataAccessException{
+	@DisplayName("Finding teacher by id")
+	void testFindTeacherById() {
 
-		Collection<Teacher> res;
+		Teacher teacher = this.teacherService.findTeacherById(1);
+		System.out.println(teacher.getFirstName());
+		assertThat(teacher.getFirstName()).startsWith("Julian");
+	}
+	//Test positivo 
+		@Test
+		@DisplayName("Finding teachers by student id")
+		void testFindTeacherByStudentId() {
+			
+			Collection<Teacher> ratedTeachers = this.teacherService.findTeacherByStudentId(1);
+			Integer size = ratedTeachers.size();
+			
+			assertTrue(size != null);
+			assertTrue(ratedTeachers!= null);
+		}
+		
+		//Test negativo
+		@Test
+		@DisplayName("Finding teachers by bad student id")
 
-		when(this.teacherRepo.findAll()).thenReturn(Lists.list(t1, t2));
+		void testFindTeachersByBadStudentId() {
+			Integer wrongStudentId = 1991;
+			assertThrows(AssertionError.class, () -> this.teacherService.findTeacherByStudentId(wrongStudentId));
+		}
 
-		res = this.teacherService.findTeachers();
+	@Test
+	@DisplayName("Finding teacher by firstName")
+	void testFindTeacherByFirstName() {
 
-		assertTrue(!res.isEmpty());
-		assertTrue(res.size()==2);
-
+		List<Teacher> teacher = this.teacherService.findTeacherByFirstName("Julian");
+		assertThat(teacher.size() == 1);
+		assertThat(teacher.get(0).getId() == 1);
 	}
 
 	@Test
-	@DisplayName("Find by id returns teacher")
-	void testFindbyId() throws DataAccessException {
-		int id = 8;
-		t1.setId(id);
-		when(teacherRepo.findById(8)).thenReturn(t1);
-		
-		ArrayList<Department> deps = (ArrayList<Department>) t1.getDepartments();
-		Department d1 = deps.get(0);
-		
-		Teacher res = teacherService.findTeacherById(id);
-		
-		assertTrue(res.getName()==t1.getName());
-		assertTrue(d1.getName().equals("Math"));
+	@DisplayName("Finding Teacher With Score")
+	void testShowTeacherWithScore() {
+
+		List<Teacher> teacher = (List<Teacher>) this.teacherService.showTeacherWithScore();
+		assertThat(teacher.size() == 1);
+		assertThat(teacher.get(0).getId() == 1);
 	}
-	
+
+	@Test
+	@DisplayName("Finding Scores by Teacher Id")
+	void testfindBySubject() {
+		int id = 1;
+		List<Score> scores = (List<Score>) this.teacherService.findScoresByTeacherId(id);
+		assertThat(scores.size() == 1);
+		assertThat(scores.get(0).getId() == 1);
+	}
+
+	@Test
+	@DisplayName("Finding Teacher by Department")
+	void findTeachersFromDepartment() {
+		int id = 5;
+		Department d = departmentService.findDepartmentById(id);
+		List<Teacher> teachers = (List<Teacher>) teacherService.findTeachers();
+		assertThat(teachers.size() == 1);
+		List<Teacher> result = new ArrayList<Teacher>();
+		for(Teacher t2 : teachers) {
+			if(t2.getDepartments().contains(d)) 
+				result.add(t2);
+		}
+		assertThat(result.size() == 1);
+		assertThat(result.get(0).getId() == 1);
+
+	}
+	// Negative test
+
 	@Test
 	@DisplayName("Finding a Teacher by bad id")
-	void testFindTeacherByBadId2(){
+	void testFindTeacherByBadId() {
 		int badId = 234234;
-		assertThrows(AssertionError.class,()->this.teacherService.findTeacherById(badId));
+		assertThrows(AssertionError.class, () -> this.teacherService.findTeacherById(badId));
+	}
+
+	// Negative test
+	
+ 	@Test
+	@DisplayName("Finding teacher by student id")
+	void testFindByStudentId(){
+		
+		List<Teacher> teacher = (List<Teacher>)this.teacherService.findTeacherByStudentId(1);
+		
+		assertThat(teacher.get(0).getFirstName()).startsWith("Julian");
+		assertThat(teacher.get(1).getFirstName()).startsWith("Popench");
+	}
+	
+	//Casos positivos
+	@Test
+	@DisplayName("Finding all teachers")
+	void testFindAll2() {
+
+		Collection<Teacher> res = this.teacherService.findTeachers();
+
+		assertTrue(!res.isEmpty());
+
 	}
 }
+
